requirements.txt
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
configparser==7.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
lxml==6.0.2
python-docx==1.2.0
python-telegram-bot[webhooks]==22.5
psycopg[binary]==3.2.10
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.6.2
